Move data processing to make_mtn_df
